feat: add Collect to token page; refactor: collect flow select token can be skipped
